+++ include/formatters/only_msg_formatter.hpp
@@ -14,4 +14,3 @@ public:
 };
 
 } // logger
-
